src/requirements.txt
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
docutils==0.16
Flask==2.0.1
future==0.18.2
google-auth==1.33.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==23.0.0
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
packaging==20.9
pandas==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
TableauScraper==0.1.19
tqdm==4.59.0
twine==3.3.0
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.4.1
